Dapper DB implementation [WIP]
Migrations now work

+++ Core/Data/Models/APIAccountRole.cs
@@ -10,12 +10,12 @@ namespace TDSM.Core.Data.Models
         /// <summary>
         /// Gets or sets the identifier.
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// Gets or sets the account identifier.
         /// </summary>
-        public int AccountId { get; set; }
+        public long AccountId { get; set; }
 
         /// <summary>
         /// Gets or sets the claim type.
