+++ Core/Data/Models/GroupNode.cs
@@ -10,17 +10,17 @@ namespace TDSM.Core.Data.Models
         /// <summary>
         /// Gets or sets the identifier.
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// Gets or sets the group identifier.
         /// </summary>
-        public int GroupId { get; set; }
+        public long GroupId { get; set; }
 
         /// <summary>
         /// Gets or sets the node identifier.
         /// </summary>
-        public int NodeId { get; set; }
+        public long NodeId { get; set; }
     }
 }
